+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@lycolia/ts-boilerplate-generator-cli",
-  "version": "0.5.6-3",
+  "version": "0.5.6-4",
   "description": "ts-server-boilerplate project Generator CLI",
   "keywords": [
     "typescript",
