+++ src/libs/ProjectCreator.ts
@@ -19,7 +19,11 @@ import {
   renameDirectory,
 } from './systems/FileSystem';
 import * as git from './systems/Git';
-import { getNpmMajorVersion, isNpmVersion7OrLater } from './systems/Npm';
+import {
+  getNpmMajorVersion,
+  installNpmModules,
+  isNpmVersion7OrLater,
+} from './systems/Npm';
 
 /**
  * create project
@@ -149,27 +153,4 @@ export const replacePackageJson = (
     .replace(/"author": .+/, `"author": "${projectOpt.author}",`)
     .replace(/"license": .+/, `"license": "${projectOpt.license}",`)
     .replace(/"url": .+/, '"url": ""');
-};
-
-/**
- * install npm modules
- * @param projectDest
- * @throws {TsgException}
- */
-export const installNpmModules = (projectDest: string) => {
-  try {
-    infoLog('Installing npm modules...');
-    const npmVer = getNpmMajorVersion();
-    if (isNpmVersion7OrLater(npmVer)) {
-      execSync(`cd ${projectDest} && npm ci --legacy-peer-deps`, {
-        stdio: 'ignore',
-      });
-    } else {
-      execSync(`cd ${projectDest} && npm ci`, {
-        stdio: 'ignore',
-      });
-    }
-  } catch (error) {
-    throw new TsgException(ErrorReasons.failNpmInst, error);
-  }
 };
