+++ src/libs/systems/specs/Npm.spec.ts
@@ -1,5 +1,12 @@
 import { execSync } from 'child_process';
-import { getNpmMajorVersion, isNpmVersion7OrLater } from '../Npm';
+import {
+  getNpmMajorVersion,
+  installNpmModules,
+  isNpmVersion7OrLater,
+} from '../Npm';
+
+jest.mock('child_process');
+const mockExecSync = execSync as jest.MockedFunction<typeof execSync>;
 
 describe('getNpmMajorVersion', () => {
   it('get number of head', () => {
@@ -15,5 +22,28 @@ describe('isNpmVersion7OrLater', () => {
   });
   it('6 or before', () => {
     expect(isNpmVersion7OrLater(6)).toBe(false);
+  });
+});
+
+describe('installNpmModules', () => {
+  mockExecSync.mockImplementation((command: string) => Buffer.from(command));
+
+  it('check npm version', () => {
+    const npmVer = getNpmMajorVersion();
+    installNpmModules('test');
+    expect(mockExecSync).toHaveBeenCalled();
+    if (isNpmVersion7OrLater(npmVer)) {
+      expect(mockExecSync).toHaveBeenNthCalledWith(
+        3,
+        'cd test && npm ci --legacy-peer-deps',
+        {
+          stdio: 'ignore',
+        }
+      );
+    } else {
+      expect(mockExecSync).toHaveBeenNthCalledWith(3, 'cd test && npm ci', {
+        stdio: 'ignore',
+      });
+    }
   });
 });
